Fix build warnings.

+++ AZ3166/src/cores/arduino/azure-iot-sdk-c/iothub_client/src/iothub_client_core_ll.c
@@ -2255,7 +2255,7 @@ IOTHUB_CLIENT_RESULT IoTHubClientCore_LL_SetOption(IOTHUB_CLIENT_CORE_LL_HANDLE 
             if (percentage > 100)
             {
                 /*Codes_SRS_IOTHUBCLIENT_LL_10_036: [Calling IoTHubClientCore_LL_SetOption with value > 100 shall return `IOTHUB_CLIENT_ERRROR`. ]*/
-                LogError("The value of diag_sampling_percentage is out of range [0, 100]: %u", percentage);
+                LogError("The value of diag_sampling_percentage is out of range [0, 100]: %u", (unsigned int)percentage);
                 result = IOTHUB_CLIENT_ERROR;
             }
             else
